release: version 1.6.0
Security and robustness:
- Validate image-parsed lengths before use (bkRead.c): directory record
  length underflow, Rock Ridge SU field bounds, short filename index,
  empty-string strip; path building now uses snprintf with bounds
- Replace bundled iniparser 4.1 (CVE-2025-0633, CVE-2023-33461) with the
  system iniparser >= 4.2.4
- Fix leaks: deleteRegFileContents freed the wrong field (every added
  file leaked its path), hard-link table pathAndName never freed,
  extractFile fd leak on error paths; ASan now reports zero leaks
- bk_set_vol_name/bk_set_publisher always NUL-terminate and reject
  overlong names; fix signed-overflow UB in El Torito checksum; use
  gmtime_r; remove unconditional debug printf output

UI and functionality:
- Edit file now writes changes back into the ISO on editor exit, using
  a unique temp directory per edit
- Long ISO operations run on a background thread with a real progress
  bar (previously the main thread blocked and the bar never updated);
  window close cancels and joins the worker before freeing C state
- Split the single-file Vala app into modules: settings.vala,
  file-item.vala, iso-operations.vala, util.vala

Packaging and CI:
- Makefile install ships desktop entry, man page and hicolor icon;
  working uninstall target; desktop Icon=isomaster (no hardcoded path)
- CI runs unit tests plus an AddressSanitizer pass, cppcheck
  informational scan; release.sh rewritten for git archive
- Version is a single source (Makefile.vala) shown in the About dialog
- 27 unit tests (3 new regression tests for malformed inputs)

+++ bk/bkWrite.c
@@ -613,7 +613,10 @@ int countTreeHeight(const DirToWrite* dir, int heightSoFar)
 * */
 unsigned short elToritoChecksum(const unsigned char* record)
 {
-    short sum;
+    /* accumulate in unsigned int: a signed 16-bit sum overflows (UB)
+    * once the words exceed 0x7FFF; the low 16 bits of the result are
+    * what the spec requires */
+    unsigned int sum;
     int i;
     
     sum = 0;
@@ -622,7 +625,7 @@ unsigned short elToritoChecksum(const unsigned char* record)
         sum += *(record + i) | (*(record + i + 1) << 8);
     }
     
-    return 0xFFFF - sum + 1;
+    return (unsigned short)(0xFFFF - sum + 1);
 }
 
 /******************************************************************************
